Move endpoint dependencies into separate modules
